Organizando aplicação para trabalhar com Chakra UI
